add template linkage/visibility and guard referenced question deletes

+++ interviewer_backend/api/routes/question.py
@@ -3,8 +3,8 @@ from typing import List, Optional
 from fastapi import APIRouter, Depends
 from fastapi_sqlalchemy import db
 
-from api.exceptions import ObjectNotFound
-from api.models.db import Question, Template
+from api.exceptions import ObjectInUse, ObjectNotFound
+from api.models.db import Question, SessionComponent, Template
 from api.schemas.models import QuestionCreate, QuestionGet, StatusResponse
 from api.utils.security import Auth, AuthUser, CsrfProtect
 
@@ -21,13 +21,20 @@ def create_question(
     db_template: Optional[Template] = Template.query(session=db.session).get(payload.template_id)
 
     if not db_template:
+        db.session.rollback()
         raise ObjectNotFound(Template, payload.template_id)
 
     new_question: Question = Question(**payload.model_dump())
     db.session.add(new_question)
+    db.session.flush()
+    response = QuestionGet(
+        id=new_question.id,
+        question=new_question.question,
+        template_id=new_question.template_id,
+    )
     db.session.commit()
 
-    return QuestionGet.model_validate(new_question)
+    return response
 
 
 @question.get("/template/{template_id}", response_model=list[QuestionGet])
@@ -38,11 +45,13 @@ def get_questions_for_template(
     db_template: Optional[Template] = Template.query(session=db.session).get(template_id)
 
     if not db_template:
+        db.session.rollback()
         raise ObjectNotFound(Template, template_id)
 
     questions: List[Question] = Question.query(session=db.session).filter(Question.template_id == template_id).all()
-
-    return [QuestionGet.model_validate(q) for q in questions]
+    response = [QuestionGet.model_validate(q) for q in questions]
+    db.session.rollback()
+    return response
 
 
 @question.delete("/{question_id}")
@@ -54,7 +63,15 @@ def delete_question(
     db_question: Optional[Question] = Question.query(session=db.session).get(question_id)
 
     if not db_question:
+        db.session.rollback()
         raise ObjectNotFound(Question, question_id)
+
+    linked_session_component = (
+        SessionComponent.query(session=db.session).filter(SessionComponent.question_id == question_id).first()
+    )
+    if linked_session_component is not None:
+        db.session.rollback()
+        raise ObjectInUse(Question, question_id, SessionComponent)
 
     db.session.delete(db_question)
     db.session.commit()
